requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg>=0.31.0
aiosqlite==0.20.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.10.0
pydantic-settings==2.5.2
pydantic[email]
slowapi==0.1.9
python-json-logger==2.0.7
reportlab==4.2.0
httpx==0.27.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Error Tracking
sentry-sdk[fastapi]==2.19.2

# AI / OpenAI
openai>=1.0.0

# Twilio (SMS + WhatsApp)
twilio>=8.0.0

# Form data handling (required for FastAPI form endpoints)
python-multipart>=0.0.6
